requirements.txt
streamlit
PyMuPDF
Pillow
google
pyperclip
google-generativeai